// javascript/lib/types/CreateServerRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DcsSettingsPayload } from "./DcsSettingsPayload";
import type { Terrain } from "./Terrain";

export type CreateServerRequest = {
  product_id: string;
  settings: DcsSettingsPayload;
  active_mods: Array<string>;
  wanted_terrains: Array<Terrain>;
};