// javascript/lib/types/NodeGame.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DcsSettings } from "./DcsSettings";
import type { NodeGameStatus } from "./NodeGameStatus";
import type { Terrain } from "./Terrain";

export type NodeGame = {
  id: string;
  node_id: string;
  user_id: string;
  product_id: string;
  ip: string;
  port: number;
  webgui_port: number;
  ftp_port: number;
  ftp_username: string;
  ftp_password: string;
  pid: number | null;
  status: NodeGameStatus;
  want_delete: boolean;
  wanted_terrains: Array<Terrain>;
  rented_at: bigint;
  active_mods: Array<string>;
  created_at: string;
  dcs_settings: DcsSettings | null;
};